+++ lib/webview_flutter_web.dart
@@ -1,5 +1,6 @@
-
-library bootpay_webview_flutter_web;
+// Copyright 2013 The Flutter Authors. All rights reserved.
+// Use of this source code is governed by a BSD-style license that can be
+// found in the LICENSE file.
 
 export 'src/http_request_factory.dart';
 export 'src/web_webview_controller.dart';