+++ test/web_webview_controller_test.dart
@@ -2,24 +2,22 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-import 'dart:html';
-// TODO(a14n): remove this import once Flutter 3.1 or later reaches stable (including flutter/flutter#104231)
-// ignore: unnecessary_import
-import 'dart:typed_data';
+import 'dart:convert';
+import 'dart:js_interop';
 
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:mockito/annotations.dart';
 import 'package:mockito/mockito.dart';
+import 'package:web/web.dart' as web;
 import 'package:bootpay_webview_flutter_platform_interface/bootpay_webview_flutter_platform_interface.dart';
 import 'package:bootpay_webview_flutter_web/webview_flutter_web.dart';
 
 import 'web_webview_controller_test.mocks.dart';
 
-@GenerateMocks(<Type>[
-  HttpRequest,
-  HttpRequestFactory,
+@GenerateMocks(<Type>[], customMocks: <MockSpec<Object>>[
+  MockSpec<HttpRequestFactory>(onMissingStub: OnMissingStub.returnDefault),
 ])
 void main() {
   WidgetsFlutterBinding.ensureInitialized();
@@ -31,8 +29,8 @@ void main() {
             WebWebViewControllerCreationParams();
 
         expect(params.iFrame.id, contains('webView'));
-        expect(params.iFrame.width, '100%');
-        expect(params.iFrame.height, '100%');
+        expect(params.iFrame.style.width, '100%');
+        expect(params.iFrame.style.height, '100%');
         expect(params.iFrame.style.border, 'none');
       });
     });
@@ -62,7 +60,7 @@ void main() {
     });
 
     group('loadRequest', () {
-      test('loadRequest throws ArgumentError on missing scheme', () async {
+      test('throws ArgumentError on missing scheme', () async {
         final WebWebViewController controller =
             WebWebViewController(WebWebViewControllerCreationParams());
 
@@ -73,8 +71,7 @@ void main() {
             throwsA(const TypeMatcher<ArgumentError>()));
       });
 
-      test('loadRequest makes request and loads response into iframe',
-          () async {
+      test('skips XHR for simple GETs (no headers, no data)', () async {
         final MockHttpRequestFactory mockHttpRequestFactory =
             MockHttpRequestFactory();
         final WebWebViewController controller =
@@ -82,17 +79,46 @@ void main() {
           httpRequestFactory: mockHttpRequestFactory,
         ));
 
-        final MockHttpRequest mockHttpRequest = MockHttpRequest();
-        when(mockHttpRequest.getResponseHeader('content-type'))
-            .thenReturn('text/plain');
-        when(mockHttpRequest.responseText).thenReturn('test data');
+        when(mockHttpRequestFactory.request(
+          any,
+          method: anyNamed('method'),
+          requestHeaders: anyNamed('requestHeaders'),
+          sendData: anyNamed('sendData'),
+        )).thenThrow(
+            StateError('The `request` method should not have been called.'));
+
+        await controller.loadRequest(LoadRequestParams(
+          uri: Uri.parse('https://flutter.dev'),
+        ));
+
+        expect(
+          (controller.params as WebWebViewControllerCreationParams).iFrame.src,
+          'https://flutter.dev/',
+        );
+      });
+
+      test('makes request and loads response into iframe', () async {
+        final MockHttpRequestFactory mockHttpRequestFactory =
+            MockHttpRequestFactory();
+        final WebWebViewController controller =
+            WebWebViewController(WebWebViewControllerCreationParams(
+          httpRequestFactory: mockHttpRequestFactory,
+        ));
+
+        final web.Response fakeResponse = web.Response(
+            'test data'.toJS,
+            <String, Object>{
+              'headers': <String, Object>{
+                'content-type': 'text/plain',
+              },
+            }.jsify()! as web.ResponseInit);
 
         when(mockHttpRequestFactory.request(
           any,
           method: anyNamed('method'),
           requestHeaders: anyNamed('requestHeaders'),
           sendData: anyNamed('sendData'),
-        )).thenAnswer((_) => Future<HttpRequest>.value(mockHttpRequest));
+        )).thenAnswer((_) => Future<Object>.value(fakeResponse));
 
         await controller.loadRequest(LoadRequestParams(
           uri: Uri.parse('https://flutter.dev'),
@@ -114,7 +140,7 @@ void main() {
         );
       });
 
-      test('loadRequest escapes "#" correctly', () async {
+      test('parses content-type response header correctly', () async {
         final MockHttpRequestFactory mockHttpRequestFactory =
             MockHttpRequestFactory();
         final WebWebViewController controller =
@@ -122,16 +148,56 @@ void main() {
           httpRequestFactory: mockHttpRequestFactory,
         ));
 
-        final MockHttpRequest mockHttpRequest = MockHttpRequest();
-        when(mockHttpRequest.getResponseHeader('content-type'))
-            .thenReturn('text/html');
-        when(mockHttpRequest.responseText).thenReturn('#');
+        final Encoding iso = Encoding.getByName('latin1')!;
+
+        final web.Response fakeResponse = web.Response(
+            String.fromCharCodes(iso.encode('España')).toJS,
+            <String, Object>{
+              'headers': <String, Object>{
+                'content-type': 'Text/HTmL; charset=latin1',
+              },
+            }.jsify()! as web.ResponseInit);
+
         when(mockHttpRequestFactory.request(
           any,
           method: anyNamed('method'),
           requestHeaders: anyNamed('requestHeaders'),
           sendData: anyNamed('sendData'),
-        )).thenAnswer((_) => Future<HttpRequest>.value(mockHttpRequest));
+        )).thenAnswer((_) => Future<Object>.value(fakeResponse));
+
+        await controller.loadRequest(LoadRequestParams(
+          uri: Uri.parse('https://flutter.dev'),
+          method: LoadRequestMethod.post,
+        ));
+
+        expect(
+          (controller.params as WebWebViewControllerCreationParams).iFrame.src,
+          'data:text/html;charset=iso-8859-1,Espa%F1a',
+        );
+      });
+
+      test('escapes "#" correctly', () async {
+        final MockHttpRequestFactory mockHttpRequestFactory =
+            MockHttpRequestFactory();
+        final WebWebViewController controller =
+            WebWebViewController(WebWebViewControllerCreationParams(
+          httpRequestFactory: mockHttpRequestFactory,
+        ));
+
+        final web.Response fakeResponse = web.Response(
+            '#'.toJS,
+            <String, Object>{
+              'headers': <String, Object>{
+                'content-type': 'text/html',
+              },
+            }.jsify()! as web.ResponseInit);
+
+        when(mockHttpRequestFactory.request(
+          any,
+          method: anyNamed('method'),
+          requestHeaders: anyNamed('requestHeaders'),
+          sendData: anyNamed('sendData'),
+        )).thenAnswer((_) => Future<Object>.value(fakeResponse));
 
         await controller.loadRequest(LoadRequestParams(
           uri: Uri.parse('https://flutter.dev'),